+++ test/katex-spec.js
@@ -1,5 +1,4 @@
 /* eslint max-len:0 */
-/* global beforeEach: false */
 /* global expect: false */
 /* global it: false */
 /* global describe: false */
@@ -7,260 +6,19 @@
 import buildMathML from "../src/buildMathML";
 import buildTree from "../src/buildTree";
 import katex from "../katex";
-import ParseError from "../src/ParseError";
 import parseTree from "../src/parseTree";
 import Options from "../src/Options";
 import Settings from "../src/Settings";
 import Style from "../src/Style";
+import {
+    defaultSettings,
+    _getBuilt, getBuilt, getParsed, stripPositions
+} from "./helpers";
 
-const defaultSettings = new Settings({
-    strict: false, // deal with warnings only when desired
-});
 const defaultOptions = new Options({
     style: Style.TEXT,
     size: 5,
     maxSize: Infinity,
-});
-
-const _getBuilt = function(expr, settings) {
-    const usedSettings = settings ? settings : defaultSettings;
-    const rootNode = katex.__renderToDomTree(expr, usedSettings);
-
-    if (rootNode.classes.indexOf('katex-error') >= 0) {
-        return rootNode;
-    }
-
-    // grab the root node of the HTML rendering
-    const builtHTML = rootNode.children[1];
-
-    // combine the non-strut children of all base spans
-    const children = [];
-    for (let i = 0; i < builtHTML.children.length; i++) {
-        children.push(...builtHTML.children[i].children.filter(
-            (node) => node.classes.indexOf("strut") < 0));
-    }
-    return children;
-};
-
-/**
- * Return the root node of the rendered HTML.
- * @param expr
- * @param settings
- * @returns {Object}
- */
-const getBuilt = function(expr, settings) {
-    const usedSettings = settings ? settings : defaultSettings;
-    expect(expr).toBuild(usedSettings);
-    return _getBuilt(expr, settings);
-};
-
-/**
- * Return the root node of the parse tree.
- * @param expr
- * @param settings
- * @returns {Object}
- */
-const getParsed = function(expr, settings) {
-    const usedSettings = settings ? settings : defaultSettings;
-
-    expect(expr).toParse(usedSettings);
-    return parseTree(expr, usedSettings);
-};
-
-const stripPositions = function(expr) {
-    if (typeof expr !== "object" || expr === null) {
-        return expr;
-    }
-    if (expr.loc && expr.loc.lexer && typeof expr.loc.start === "number") {
-        delete expr.loc;
-    }
-    Object.keys(expr).forEach(function(key) {
-        stripPositions(expr[key]);
-    });
-    return expr;
-};
-
-const parseAndSetResult = function(expr, result, settings) {
-    try {
-        return parseTree(expr, settings || defaultSettings);
-    } catch (e) {
-        result.pass = false;
-        if (e instanceof ParseError) {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with error: " + e.message;
-        } else {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with unknown error: " + e.message;
-        }
-    }
-};
-
-const buildAndSetResult = function(expr, result, settings) {
-    try {
-        return _getBuilt(expr, settings || defaultSettings);
-    } catch (e) {
-        result.pass = false;
-        if (e instanceof ParseError) {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with error: " + e.message;
-        } else {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with unknown error: " + e.message;
-        }
-    }
-};
-
-beforeEach(function() {
-    expect.extend({
-        toParse: function(actual, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: true,
-                message: () => "'" + actual + "' succeeded parsing",
-            };
-            parseAndSetResult(actual, result, usedSettings);
-            return result;
-        },
-
-        toNotParse: function(actual, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: false,
-                message: () => "Expected '" + actual + "' to fail " +
-                    "parsing, but it succeeded",
-            };
-
-            try {
-                parseTree(actual, usedSettings);
-            } catch (e) {
-                if (e instanceof ParseError) {
-                    result.pass = true;
-                    result.message = () => "'" + actual + "' correctly " +
-                        "didn't parse with error: " + e.message;
-                } else {
-                    result.message = () => "'" + actual + "' failed " +
-                        "parsing with unknown error: " + e.message;
-                }
-            }
-
-            return result;
-        },
-
-        toBuild: function(actual, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: true,
-                message: () => "'" + actual + "' succeeded in building",
-            };
-
-            expect(actual).toParse(usedSettings);
-
-            try {
-                _getBuilt(actual, usedSettings);
-            } catch (e) {
-                result.pass = false;
-                if (e instanceof ParseError) {
-                    result.message = () => "'" + actual + "' failed to " +
-                        "build with error: " + e.message;
-                } else {
-                    result.message = () => "'" + actual + "' failed " +
-                        "building with unknown error: " + e.message;
-                }
-            }
-
-            return result;
-        },
-
-        toNotBuild: function(actual, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: false,
-                message: () => "Expected '" + actual + "' to fail " +
-                    "building, but it succeeded",
-            };
-
-            try {
-                _getBuilt(actual, usedSettings);
-            } catch (e) {
-                if (e instanceof ParseError) {
-                    result.pass = true;
-                    result.message = () => "'" + actual + "' correctly " +
-                        "didn't build with error: " + e.message;
-                } else {
-                    result.message = () => "'" + actual + "' failed " +
-                        "building with unknown error: " + e.message;
-                }
-            }
-
-            return result;
-        },
-
-        toParseLike: function(actual, expected, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: true,
-                message: () => "Parse trees of '" + actual +
-                    "' and '" + expected + "' are equivalent",
-            };
-
-            const actualTree = parseAndSetResult(actual, result,
-                usedSettings);
-            if (!actualTree) {
-                return result;
-            }
-            const expectedTree = parseAndSetResult(expected, result,
-                usedSettings);
-            if (!expectedTree) {
-                return result;
-            }
-
-            stripPositions(actualTree);
-            stripPositions(expectedTree);
-
-            if (JSON.stringify(actualTree) !== JSON.stringify(expectedTree)) {
-                result.pass = false;
-                result.message = () => "Parse trees of '" + actual +
-                    "' and '" + expected + "' are not equivalent";
-            }
-            return result;
-        },
-
-        toBuildLike: function(actual, expected, settings) {
-            const usedSettings = settings ? settings : defaultSettings;
-
-            const result = {
-                pass: true,
-                message: () => "Build trees of '" + actual +
-                    "' and '" + expected + "' are equivalent",
-            };
-
-            const actualTree = buildAndSetResult(actual, result,
-                usedSettings);
-            if (!actualTree) {
-                return result;
-            }
-            const expectedTree = buildAndSetResult(expected, result,
-                usedSettings);
-            if (!expectedTree) {
-                return result;
-            }
-
-            stripPositions(actualTree);
-            stripPositions(expectedTree);
-
-            if (JSON.stringify(actualTree) !== JSON.stringify(expectedTree)) {
-                result.pass = false;
-                result.message = () => "Parse trees of '" + actual +
-                    "' and '" + expected + "' are not equivalent";
-            }
-            return result;
-        },
-    });
 });
 
 describe("A parser", function() {
