+++ test/unicode-spec.js
@@ -7,69 +7,9 @@ import ParseError from "../src/ParseError";
 import parseTree from "../src/parseTree";
 import Settings from "../src/Settings";
 import {scriptFromCodepoint, supportedCodepoint} from "../src/unicodeScripts";
-
-const defaultSettings = new Settings({
-    strict: false, // deal with warnings only when desired
-});
-const strictSettings = new Settings({strict: true});
-
-const parseAndSetResult = function(expr, result, settings) {
-    try {
-        return parseTree(expr, settings || defaultSettings);
-    } catch (e) {
-        result.pass = false;
-        if (e instanceof ParseError) {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with error: " + e.message;
-        } else {
-            result.message = () => "'" + expr + "' failed " +
-                "parsing with unknown error: " + e.message;
-        }
-    }
-};
+import {defaultSettings, strictSettings, parseAndSetResult} from "./helpers";
 
 describe("unicode", function() {
-    beforeEach(function() {
-        expect.extend({
-
-            toParse: function(actual, settings) {
-                const usedSettings = settings ? settings : defaultSettings;
-
-                const result = {
-                    pass: true,
-                    message: () => "'" + actual + "' succeeded parsing",
-                };
-                parseAndSetResult(actual, result, usedSettings);
-                return result;
-            },
-
-            toNotParse: function(actual, settings) {
-                const usedSettings = settings ? settings : defaultSettings;
-
-                const result = {
-                    pass: false,
-                    message: () => "Expected '" + actual + "' to fail " +
-                        "parsing, but it succeeded",
-                };
-
-                try {
-                    parseTree(actual, usedSettings);
-                } catch (e) {
-                    if (e instanceof ParseError) {
-                        result.pass = true;
-                        result.message = () => "'" + actual + "' correctly " +
-                            "didn't parse with error: " + e.message;
-                    } else {
-                        result.message = () => "'" + actual + "' failed " +
-                            "parsing with unknown error: " + e.message;
-                    }
-                }
-
-                return result;
-            },
-        });
-    });
-
     it("should parse Latin-1 inside \\text{}", function() {
         expect('\\text{ÀÁÂÃÄÅÈÉÊËÌÍÎÏÑÒÓÔÕÖÙÚÛÜÝàáâãäåèéêëìíîïñòóôõöùúûüýÿ' +
             'ÆÇÐØÞßæçðøþ}').toParse();
